PLP/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
gunicorn==23.0.0
sqlalchemy==2.0.36
aiomysql==0.2.0
alembic==1.13.3
pydantic-settings==2.6.1
email-validator==2.2.0
python-multipart==0.0.12
aiofiles==24.1.0
PyJWT[crypto]==2.9.0
httpx==0.27.2
celery[redis]==5.4.0
redis==5.1.1
slowapi==0.1.9
openai==1.51.0
faster-whisper==1.0.3
python-dotenv==1.0.1
